+++ src/types/ethers-types.ts
@@ -1,0 +1,13 @@
+export type {
+  Block,
+  BlockTag,
+  BlockWithTransactions,
+  EventType,
+  FeeData,
+  Log,
+  TransactionRequest,
+  TransactionResponse,
+  TransactionReceipt
+} from '@ethersproject/abstract-provider';
+
+export type { BigNumberish } from '@ethersproject/bignumber';