+++ test/integration/nft.test.ts
@@ -124,6 +124,7 @@ describe('E2E integration tests', () => {
         .length
     ).toBeGreaterThan(0);
     expect(response.ownedNfts.length).toEqual(51);
+    expect(typeof response.blockHash).toEqual('string');
   });
 
   it('getOwnersForNft() from NFT', async () => {
@@ -199,11 +200,14 @@ describe('E2E integration tests', () => {
     verifyNftContractMetadata(response.contracts[0]);
   });
 
-  it('getContractsForOwner() with pageKey', async () => {
-    const firstPage = await alchemy.nft.getContractsForOwner(ownerEns);
+  it('getContractsForOwner() with pageKey and pageSize', async () => {
+    const firstPage = await alchemy.nft.getContractsForOwner(ownerEns, {
+      pageSize: 4
+    });
 
     expect(firstPage.pageKey).toBeDefined();
     expect(typeof firstPage.pageKey).toEqual('string');
+    expect(firstPage.contracts.length).toEqual(4);
 
     const response = await alchemy.nft.getContractsForOwner(ownerEns, {
       pageKey: firstPage?.pageKey
@@ -269,6 +273,19 @@ describe('E2E integration tests', () => {
           nft.contract.totalSupply !== undefined
       ).length
     ).toBeGreaterThan(0);
+  });
+
+  it('getContractMetadataBatch()', async () => {
+    const contractAddresses = [
+      '0xe785e82358879f061bc3dcac6f0444462d4b5330',
+      '0xbc4ca0eda7647a8ab7c2061c2e118a18a936f13d'
+    ];
+    const response = await alchemy.nft.getContractMetadataBatch(
+      contractAddresses
+    );
+    expect(response.length).toEqual(2);
+    expect(contractAddresses.includes(response[0].address)).toEqual(true);
+    expect(contractAddresses.includes(response[1].address)).toEqual(true);
   });
 
   it('getNftsForOwnerIterator()', async () => {